[veSync] PMD Error correction - file naming correction
