--- bundles/org.openhab.binding.vesync/src/main/java/org/openhab/binding/vesync/internal/dto/responses/VesyncManagedDevicesPage.java
@@ -1,178 +1,0 @@
-/**
- * Copyright (c) 2010-2022 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.vesync.internal.dto.responses;
-
-import com.google.gson.annotations.SerializedName;
-
-/**
- * The {@link VesyncManagedDevicesPage} is a Java class used as a DTO to hold the Vesync's API's response data to a
- * page of data requesting the manages devices.
- *
- * @author David Goodyear - Initial contribution
- */
-public class VesyncManagedDevicesPage extends VesyncResponse {
-
-    @SerializedName("result")
-    public Result result;
-
-    public class Result {
-        @SerializedName("pageNo")
-        public String pageNo;
-
-        @SerializedName("total")
-        public String total;
-
-        @SerializedName("pageSize")
-        public String pageSize;
-
-        @SerializedName("list")
-        public VesyncManagedDeviceBase[] list;
-
-        /**
-         * Contains basic information about the device.
-         */
-        public class VesyncManagedDeviceBase {
-
-            @SerializedName("deviceRegion")
-            public String deviceRegion;
-
-            public String getDeviceRegion() {
-                return deviceRegion;
-            }
-
-            @SerializedName("deviceType")
-            public String deviceType;
-
-            public String getDeviceType() {
-                return deviceType;
-            }
-
-            @SerializedName("deviceName")
-            public String deviceName;
-
-            public String getDeviceName() {
-                return deviceName;
-            }
-
-            @SerializedName("deviceImg")
-            public String deviceImg;
-
-            public String getDeviceImg() {
-                return deviceImg;
-            }
-
-            @SerializedName("deviceStatus")
-            public String deviceStatus;
-
-            public String getDeviceStatus() {
-                return deviceStatus;
-            }
-
-            @SerializedName("cid")
-            public String cid;
-
-            public String getCid() {
-                return cid;
-            }
-
-            @SerializedName("connectionStatus")
-            public String connectionStatus;
-
-            public String getConnectionStatus() {
-                return connectionStatus;
-            }
-
-            @SerializedName("connectionType")
-            public String connectionType;
-
-            public String getConnectionType() {
-                return connectionType;
-            }
-
-            @SerializedName("type")
-            public String type;
-
-            public String getType() {
-                return type;
-            }
-
-            @SerializedName("subDeviceNo")
-            public String subDeviceNo;
-
-            public String getSubDeviceNo() {
-                return subDeviceNo;
-            }
-
-            @SerializedName("subDeviceType")
-            public String subDeviceType;
-
-            public String getSubDeviceType() {
-                return subDeviceType;
-            }
-
-            @SerializedName("uuid")
-            public String uuid;
-
-            public String getUuid() {
-                return uuid;
-            }
-
-            @SerializedName("macID")
-            public String macId;
-
-            public String getMacId() {
-                return macId;
-            }
-
-            @SerializedName("currentFirmVersion")
-            public String currentFirmVersion;
-
-            public String getCurrentFirmVersion() {
-                return currentFirmVersion;
-            }
-
-            @SerializedName("configModule")
-            public String configModule;
-
-            public String getConfigModule() {
-                return configModule;
-            }
-
-            @SerializedName("mode")
-            public String mode;
-
-            public String getMode() {
-                return mode;
-            }
-
-            @SerializedName("speed")
-            public String speed;
-
-            public String getSpeed() {
-                return speed;
-            }
-        }
-
-        public String getPageNo() {
-            return pageNo;
-        }
-
-        public String getPageSize() {
-            return pageSize;
-        }
-
-        public String getTotal() {
-            return total;
-        }
-    }
-}
